Crate ldk_node

A library providing a simplified API for the Lightning Dev Kit. While LDK itself provides a highly configurable and adaptable interface, this API champions simplicity and ease of use over configurability. To this end, it provides an opionated set of design choices and ready-to-go default modules, while still enabling some configurability when dearly needed by the user:

Chain data is accessed through an Esplora client.

Wallet and channel states are persisted to disk.

Gossip is retrieved over the P2P network.

Re-exports

pub use bitcoin;

pub use lightning;

pub use lightning_invoice;

Structs

Builder

A builder for an Node instance, allowing to set some configuration and module choices from the getgo.

Config

Represents the configuration of an Node instance.

Node

The main interface object of the simplified API, wrapping the necessary LDK and BDK functionalities.

PaymentInfo

Represents a payment.

Enums

Error

An error that possibly needs to be handled by the user.

Event

An event emitted by Node, which should be handled by the user.

PaymentStatus

Represents the current status of a payment.