Enum ldk_node::Event

pub enum Event {
    PaymentSuccessful {
        payment_hash: PaymentHash,
    },
    PaymentFailed {
        payment_hash: PaymentHash,
    },
    PaymentReceived {
        payment_hash: PaymentHash,
        amount_msat: u64,
    },
    ChannelReady {
        channel_id: [u8; 32],
        user_channel_id: u128,
    },
    ChannelClosed {
        channel_id: [u8; 32],
        user_channel_id: u128,
    },
}

An event emitted by Node, which should be handled by the user.

Variants

PaymentSuccessful

Fields

payment_hash: PaymentHash

The hash of the payment.

A sent payment was successful.

PaymentFailed

Fields

payment_hash: PaymentHash

The hash of the payment.

A sent payment has failed.

PaymentReceived

Fields

payment_hash: PaymentHash

The hash of the payment.

amount_msat: u64

The value, in thousandths of a satoshi, that has been received.

A payment has been received.

ChannelReady

Fields

channel_id: [u8; 32]

The channel_id of the channel.

user_channel_id: u128

The user_channel_id of the channel.

A channel is ready to be used.

ChannelClosed

Fields

channel_id: [u8; 32]

The channel_id of the channel.

user_channel_id: u128

The user_channel_id of the channel.

A channel has been closed.