Struct ldk_node::Node

pub struct Node { /* private fields */ }

The main interface object of the simplified API, wrapping the necessary LDK and BDK functionalities.

Needs to be initialized and instantiated through Builder::build.

Implementations

impl Node

pub fn start(&self) -> Result<(), Error>

Starts the necessary background tasks, such as handling events coming from user input, LDK/BDK, and the peer-to-peer network. After this returns, the Node instance can be controlled via the provided API methods in a thread-safe manner.

pub fn stop(&self) -> Result<(), Error>

Disconnects all peers, stops all running background tasks, and shuts down Node.

pub fn next_event(&self) -> Event

Blocks until the next event is available.

Note: this will always return the same event until handling is confirmed via Node::event_handled.

pub fn event_handled(&self)

Confirm the last retrieved event handled.

pub fn node_id(&self) -> Result<PublicKey, Error>

Returns our own node id

pub fn listening_address(&self) -> Option<String>

Returns our own listening address and port.

pub fn new_funding_address(&self) -> Result<Address, Error>

Retrieve a new on-chain/funding address.

pub fn on_chain_balance(&self) -> Result<Balance, Error>

Retrieve the current on-chain balance.

pub fn list_channels(&self) -> Vec<ChannelDetails>

Retrieve a list of known channels.

pub fn connect_open_channel(
&self,
node_pubkey_and_address: &str,
channel_amount_sats: u64,
announce_channel: bool
) -> Result<(), Error>

Connect to a node and open a new channel. Disconnects and re-connects are handled automatically

Returns a temporary channel id

pub fn sync_wallets(&self) -> Result<(), Error>

Sync the LDK and BDK wallets with the current chain state.

Note that the wallets will be also synced regularly in the background.

pub fn close_channel(
&self,
channel_id: &[u8; 32],
counterparty_node_id: &PublicKey
) -> Result<(), Error>

Close a previously opened channel.

pub fn send_payment(&self, invoice: Invoice) -> Result<PaymentHash, Error>

Send a payement given an invoice.

pub fn send_spontaneous_payment(
&self,
amount_msat: u64,
node_id: &str
) -> Result<PaymentHash, Error>

Send a spontaneous, aka. “keysend”, payment

pub fn receive_payment(
&self,
amount_msat: Option<u64>,
description: &str,
expiry_secs: u32
) -> Result<Invoice, Error>

Returns a payable invoice that can be used to request and receive a payment.

pub fn payment_info(&self, payment_hash: &[u8; 32]) -> Option<PaymentInfo>

Query for information about the status of a specific payment.